CMake - Fix shared runtimes install path
Currently (well, at least on Windows), the created DLL files aren't installed to the correct directory after a build. The dll files are installed to the 'lib' directory. Which actually is the place for the static library or the shared library's symbols file (.lib).
The shared library runtimes should be installed to the 'bin' output directory, as proposed in this PR.
It also follows convention as used in other notable libraries as libpng, libogg/vorbis, freetype, zlib, tinyxml, etc.

See the ```CONFIGURATIONS``` option in [Install command introduction in the CMake manual](https://cmake.org/cmake/help/v3.14/command/install.html#introduction), .