Update CMakeLists.txt
Currently (well, at least on Windows), the created DLL files aren't installed in the correct output directory after a build. The dll files are installed in the 'lib' directory. Which actually is the place for the static library or the shared library's symbols file (.lib), not the actual shared library's runtime.
The shared library runtimes should be installed to the 'bin' output directory, as proposed in this PR.

See the ```CONFIGURATIONS``` option in [Install command introduction in the CMake manual](https://cmake.org/cmake/help/v3.14/command/install.html#introduction).

This also follows convention as used in other notable libraries as libpng, libogg/vorbis, freetype, zlib, tinyxml, etc.